// src/app/models/admin/capphephoatdongkhoangsan/cppheduyettruluongkhoangsan/cppheduyettruluongloaikhoangsan.model.ts
export class OutputCpPheDuyetTLLoaiKSModel {
    idpheduyettruluongloaikhoangsan: string;
    tenkhoangsan: string;
    idloaikhoangsan: string;
    tenloaikhoangsan: string;
    tennhomkhoangsan: string;
    idcaptruluong: string;
    truluong: number;
    donvitruluong: string;
    phanloai: number;
    thutu: number;
    idpheduyettruluong: string;
  }
  
  export class InputCpPheDuyetTLLoaiKSModel {
    idpheduyettruluongloaikhoangsan: string;
    tenkhoangsan: string;
    idloaikhoangsan: string;
    tenloaikhoangsan: string;
    tennhomkhoangsan: string;
    idcaptruluong: string;
    truluong: number;
    donvitruluong: string;
    phanloai: number;
    thutu: number;
    idpheduyettruluong: string;
  }
  